move store.js into lib folder
