+++ lib/store.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
 import { createLogger } from 'redux-logger';
-import rootReducer from './modules';
+import rootReducer from '../modules';
 
 const exempleInitialState = {};
 const enhancers = [];
@@ -22,8 +22,7 @@ const setupStore = (initialState = exempleInitialState) => {
   if (process.env.NODE_ENV !== 'production') {
     if (module.hot) {
       module.hot.accept('./modules', () => {
-        // eslint-disable-next-line global-require
-        const nextRootReducer = require('./modules');
+        const nextRootReducer = require('../modules');
         store.replaceReducer(nextRootReducer, store.state);
       });
     }
