+++ server.js
@@ -1,5 +1,4 @@
 /* eslint-disable no-console */
-
 const express = require('express');
 const next = require('next');
 require('dotenv').config();
